refactor: change test to fit with new html template

+++ src/Fetcher/fetcher.go
@@ -46,8 +46,10 @@ func FetchFiles(url, branch, specFile string) error {
 	}
 
 	for folder, _ := range seenFolders {
-		os.RemoveAll(("./" + folder))
-	}
+        if folder != "cache" {
+            os.RemoveAll(("./" + folder))
+        }
+    }
 
 	return nil
 }
